+++ dev/index.js
@@ -1,0 +1,2 @@
+const moduleName = process.argv[2] || 'action-chaining'
+module.exports = require('./' + moduleName)